# OrderPlatform/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 电商平台 API
  description: 分布式微服务架构下的RESTful API规范
  version: 1.0.0
  contact:
    name: 开发者
    email: dev@example.com
servers:
  - url: http://localhost:8080/api
    description: 本地网关
  # - url: https://api.yourdomain.com/api
  #   description: 生产环境
security:
  -BearerAuth: []
paths:
  /auth/register:
    post:
      tags: [认证]
      summary: 用户注册
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '200':
          description: 注册成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/Error'
  /auth/login:
    post:
      tags: [认证]
      summary: 用户登录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        userId:
          type: integer
        username:
          type: string
        token:
          type: string
    RegisterRequest:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
...
  responses:
    Error:
      description: 错误响应
      content:
        application/json:
          schema:
            type: object
            properties:
              code:
                type: integer
              message:
                type: string